Terminado p3 e3 creo

+++ p3_ejercicio_3/src/main/java/ar/edu/unlp/info/oo2/p3_ejercicio_3/CRC_Strategy.java
@@ -1,5 +1,5 @@
 package ar.edu.unlp.info.oo2.p3_ejercicio_3;
 
 public interface CRC_Strategy {
-	public Integer crcFor(String data);
+	public long crcFor(String data);
 }
